requirements.txt
autopep8==2.0.2
coverage==7.2.7
flake8==6.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-cov==4.1.0
python-dotenv==1.0.0
